Refactor naming.
